# Fixes for `cargo bench` and new benchmarks (end-to-end + base64 optimization).