… test_data/large_base64.in