Summary
========

- Fixing `cargo bench --features=bench`
- Fixing `bench_optimize` to also cover `Optimizer::next`
- Adding a benchmark for running `Optimizer` on base64 data.
- Adding a benchmark for `QrCode::with_error_correction_level`


Fixing `cargo bench --features=bench`
=================================

This commit makes `use std:convert::TryInto` conditional via
`#[cfg(debug_assertions)]`.  This avoids the following build error:

    ```
    $ cargo bench
    ...
    error: unused import: `std::convert::TryInto`
     --> src/cast.rs:1:5
      |
    1 | use std::convert::TryInto;
      |     ^^^^^^^^^^^^^^^^^^^^^
    ```

This commit also opts into the `test` feature (only if the "bench"
feature is requested, because benchmarking is only available in the
nightly version of the compiler) and declares the `extern crate test`
dependency on the compiler-provided `test` crate. This avoids the
following build errors:

    ```
    $ cargo bench
    ...
    error[E0433]: failed to resolve: use of undeclared crate or module `test`
       --> src/optimize.rs:710:33
        |
    710 | fn bench_optimize(bencher: &mut test::Bencher) {
        |                                 ^^^^
                   use of undeclared crate or module `test`
    ```

    ```
    $ cargo bench
    ...
    error[E0658]: use of unstable library feature 'test'
       --> src/optimize.rs:710:33
        |
    710 | fn bench_optimize(bencher: &mut test::Bencher) {
        |                                 ^^^^^^^^^^^^^
        |
        = note: see issue #50297
                <https://github.com/rust-lang/rust/issues/50297>
                for more information
        = help: add `#![feature(test)]` to the crate attributes to enable
    ```

This commit also makes small tweaks to how `README.md` is included.
It seems that this doesn't actually test the examples in the `README.md`
file (this commit adds a TODO for this observation), but it does avoid
the following build error:

    ```
    $ cargo bench --features=bench
    ...
    error: unknown `doc` attribute `include`
      --> src/lib.rs:14:36
       |
    14 | #![cfg_attr(feature = "bench", doc(include = "../README.md"))]
       |                                ----^^^^^^^^^^^^^^^^^^^^^^^^- help:
       |                                use `doc = include_str!` instead:
       |                           `#![doc = include_str!("../README.md")]`
       |
       = warning: this was previously accepted by the compiler but is being
         phased out; it will become a hard error in a future release!
       = note: for more information, see issue #82730
         <https://github.com/rust-lang/rust/issues/82730>
    ```

Finally, this commit declares the "bench" feature in `Cargo.toml`.
After these changes the following command line succeeds:

    ```
    $ cargo bench --features=bench
    ...
    test bits::bench_find_min_version    ... bench:           1 ns/iter (+/- 0)
    test bits::bench_push_splitted_bytes ... bench:       3,862 ns/iter (+/- 58)
    test optimize::bench_optimize        ... bench:          19 ns/iter (+/- 0)


Fixing `bench_optimize` to also cover `Optimizer::next`.
=============================================

Main change - fixing `bench_optimize`
-------------------------------------

Before this commit `bench_optimize` would mostly measure `Parser::new`
and `Optimizer::new`, but would *not* measure `Parser::next` nor
`Optimizer::next`.  After this commit the measurement covers more of the
parsing and optmization code.

Benchmarks before this commit:

test bits::bench_find_min_version    ... bench:     1 ns/iter (+/- 0)
test bits::bench_push_splitted_bytes ... bench: 4,158 ns/iter (+/- 57)
test optimize::bench_optimize        ... bench:     1 ns/iter (+/- 0)

Benchmarks after this commit (note the expected+desirable change
in `bench_optimize`):

test bits::bench_find_min_version    ... bench:     1 ns/iter (+/- 0)
test bits::bench_push_splitted_bytes ... bench: 4,155 ns/iter (+/- 79)
test optimize::bench_optimize        ... bench: 1,617 ns/iter (+/- 96)

Supporting change - tweaking `total_encoded_len`
------------------------------------------------

This commit changes the public API of `total_encoded_len` to accept a
broader set of arguments (i.e. this is a backcompatible change and
doesn't require changing the crate version).  Before this commit, only
`&[Segment]` was accepted, but after this commit an `Optimizer` is also
a valid argument.  This change in the public API helps reuse
`total_encoded_len` when calling it with `Optimizer` from
`bench_optimizer`.

FWIW, an earlier discussion about an idiomatic way to accept iterators
over `T` or `&T` can be found at:
https://users.rust-lang.org/t/idiomatic-way-to-return-receive-iterator/32383/5
(there, the recommendation is also to use the `Borrow` trait).