@anforowicz 2023-04-20
This PR is primarily motivated by the desire to lay down the foundation for evaluating the performance impact of switching the segmentation optimization algorithm in a future PR (see the last 2 commits in https://github.com/anforowicz/qr_code/tree/shortest-path-algorithm;  the first 4 commits are the benchmarking tweaks in this PR).

PS. This PR ended up quite big - it might make sense to land each of the commits separately.  Please shout if you'd like me to split this PR into 4 smaller PRs.