@anforowicz 2023-04-27
@RCasatta can you PTAL?

I've rebased + made some tweaks to ensure that rustc stable and 1.48 work fine:

- I had to comment out the `"bench"` feature in `Cargo.toml` because otherwise `cargo ... --all-features` would try to use this feature with rustc/stable (and benchmarking is only available on rustc/nightly).  Commenting out is a bit icky, but still seems better than detecting rustc/nightly through a build script.

- I also had to revert changes to use `include_str` - apparently this didn't work in 1.48 (not sure exactly why, but this was an opportunistic change in response to clippy, so reverting this seems fine/easiest).

PS. Sorry about the forced rebases in my fork.  I wanted to make sure that everything is green after every commit (not just after the whole PR).  But... I don't use GitHub often, so please shout if this is bad form / if I should have just appended a new commit with fixes...